Keep eglot from removing other flymake backends
Eglot, like other lsp clients, takes over management of the buffer. It
disables existing flymake backends and adds its own assuming that the
lsp server provides equivalent features. Pyright does not support
linting however and so we want to keep the existing flymake backends
(pylint and flake8) enabled.

This change sets eglot-stay-out-of to keep eglot from taking over
flymake and then adds an eglot-managed-mode-hook to manually add the
flymake backend for eglot as well.

Probably this should only be applied to python-mode, or have a way to
choose which modes are configured this way.