Match garden-exec header for python mode
This is the beginning of a way to match and parse garden-exec headers
to not only set the mode but also set up the environment according to
garden load and path operations in the header. This could be done by
using MATCH-FUNCTION in the alist instead of a regex, and adding to
magic-mode-alist instead, which has higer priority than
auto-mode-alist.