Try removing ':straight (:type build-in)'
This is not exactly a no-op. Since I'm not using
straight-use-package-by-default straight.el will not try to install
everything configured with use-package, but it will try to install
dependencies. If a dependency was configured with ':straight (:type
built-in)' then straight.el would obey and not try to install the
dependency. Without ':straight (:type built-in)' it does install the
dependency, even if it is a built-in emacs feature. Currently, e.g.,
this causes flymake to be installed from the straight.el elpa mirror
as it's a dependency of flymake-collection.